requirements-2023-pypi-cuda.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.6
cycler==0.11.0
fonttools==4.25.0
h5py==3.7.0
idna==3.4
importlib-resources==5.7.1
kiwisolver==1.4.4
matplotlib==3.5.3
munkres==1.1.4
numpy==1.19.5
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
ply==3.11
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
toml==0.10.2
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.15
zipp==3.8.1
